+++ models/student.js
@@ -3,12 +3,12 @@ module.exports = function(sequelize, DataTypes) {
     name: {
       type: DataTypes.STRING,
       allowNull: false,
-      defaultValue: "John Doe"
+      // defaultValue: "John Doe"
     },
     email: {
       type: DataTypes.STRING,
       allowNull: false,
-      defaultValue: "fakeEmail@email.com"
+      // defaultValue: "fakeEmail@email.com"
     },
     description: {
       type: DataTypes.TEXT,
@@ -16,17 +16,17 @@ module.exports = function(sequelize, DataTypes) {
     },
     rating: {
       type: DataTypes.INTEGER,
-      defaultValue: 5
+      // defaultValue: 5
     },
     image: {
       type: DataTypes.STRING,
       allowNull: false,
-      defaultValue: "https://via.placeholder.com/150"
+      // defaultValue: "https://via.placeholder.com/150"
     },
     googleId: {
       type: DataTypes.STRING,
       allowNull: false,
-      defaultValue: "fake googleId"
+      // defaultValue: "fake googleId"
     }
   });
